[INITIAL] Basic object type & GC
This is a basic object system that allows allocating objects and having
them be garbage collected. Object types should inherit from Object, and
will need to override u32 GetTypeId() const and String GetTypeName()
const at a minimum.

Objects can expose properties that can be reflected on by implementing
GetObjectFields(). Typically users will use the EXPOSE_FIELD macro to
handle populating the list of fields that are exposed to the reflection
system. Object fields contain the name, type and offset of the property,
and can provide a pointer to the field given an object instance of the
appropriate type.

Using this system, objects can expose pointers to other objects
(including arrays of other objects). We use this to implement a basic
mark & sweep garbage collector. Objects that are eligible to be garbage
collected can be allocated with NewObject<T>(). Objects allocated with
NewObject have a stable location in memory. Some objects should be
added to the root set by calling AddToRootSet() on them. They can later
be removed by using RemoveFromRootSet(). When appropriate, call the
static Object::CollectGarbage() method to run garbage collection.
Objects that are not referencable from the root set will be destroyed.

Objects can be manually destroyed by calling Destroy() on them. In this
case the object is no longer valid, and calling IsValid(object) will
return false. During the next garbage collection, and pointers to
destroyed or mid-destroy objects will be set to nullptr.

Objects can declare if they are ready to finish being destroyed or not
by implementing the IsDestroyFinished() method. If this method returns
false, then the destructor will not be called and the memory not freed
until the next garbage collection. This allows objects to imply that
they need more time before they should be fully destroyed, typically
because they need to async release a resource.

StrongObjectPtrs allow holding a reference to an object and keeping it
alive, even if it is not otherwise referencable from the root set. These
should be used sparingly, as the object memory cannot be reclaimed until
these pointers are freed, and this is typically only useful for keeping
objects alive for a short while, e.g. during an async operation.

WeakObjectPtrs allow holding a reference to an object and checking if it
is still valid without preventing garbage collection.

The test suite is accessible from CMake. Typically running make check in
the root directory is sufficient to build & run the tests. Tests use the
check2 framework.

A note on the handling of memory locations. In order to ensure that once
allocated, objects have a stable memory address, we keep pools of
objects allocated. Each pool is responsible for allocating objects of a
certain size (note that future work is to try and group similar sized
objects into the same pool. E.g. a 22 byte and 24 byte object could both
just use a 32 byte pool for example). Each pool maintains a list of
blocks. Once allocated blocks are never deallocated, nor resized. If we
wish to allocate from an object pool but do not have spare space, we
will allocate a new block and sub-allocate an object from that block in
the pool. Each object has a 16 byte header placed just before it, which
allows storing flags & a free list pointer. We also store a generation
in this header, which allows reusing block elements without causing weak
object pointers to incorrectly refer to a newly allocated object.
Effectively this causes the object pools to become a slotmap.