Add basic class object to allow inspect class heirarchies
Here we add a Class type. The Class type represents the class of an
object, including its parent class and fields on the type. These classes
are discovered during static initialization, and require that any object
types include a DECLARE_OBJECT(objectType) line in the header, and a
IMPL_OBJECT(objectType, parentType) in a single translation unit.