Remove need for initial set unreachable phase
Previously we iterated all objects setting their unreachable flag. We do
not need to do this - we can mark them as unreachable on allocation,
then if the mark phase marks them as rechable, the sweep phase can mark
them as unreachable again instead of freeing them. By doing this we only
have to iterate over the object array once during the sweep phase, and
avoid an initial mark unreachable phase.